AudioManager α版
