+++ MapMakingStage1/Assets/Art/Materials/StarsBackground.mat
@@ -89,7 +89,7 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     - __dirty: 0
-    - _sky: 0
+    - _sky: 0.044
     m_Colors:
     - _Color: {r: 1, g: 1, b: 1, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
@@ -97,5 +97,5 @@ Material:
     - _Sky2: {r: 1, g: 1, b: 1, a: 0}
     - _SpecColor: {r: 1, g: 1, b: 1, a: 1}
     - _StarsColor: {r: 1, g: 1, b: 1, a: 0}
-    - _StarsTiling: {r: 3, g: 3, b: 0, a: 0}
+    - _StarsTiling: {r: 6, g: 6, b: 0, a: 0}
     - _sss: {r: 0, g: 0, b: 0, a: 0}
