+++ MapMakingStage1/Assets/Resources/Audio/BGM.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6e9b2b07ce6e7bf40b9e9fe8813273ab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
