+++ MapMakingStage1/Assets/Resources/Audio/SE.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8fe4aff3f375b3c40bcb363da9ea2649
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
