--- MapMakingStage1/Assets/User/Suda/Editor/SceneNameCreator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9bfb14d2dc1517c449b5883647886f4f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
