+++ MapMakingStage1/Assets/User/Suda/scripts/AUDIO.cs
@@ -1,0 +1,9 @@
+﻿/// <summary>
+/// オーディオ名を定数で管理するクラス
+/// </summary>
+public static class AUDIO
+{
+	  public const string BGM_STAGE1_SAMPLE = "Stage1_Sample";
+	
+	  public const string SE_FANFARE_SAMPLE = "Fanfare_Sample";
+}
