+++ MapMakingStage1/Assets/User/Suda/scripts/AUDIO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a59cb14f6d43d0488eb5f81a4e4a5bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
