--- MapMakingStage1/Assets/User/Suda/scripts/AudioNameCreator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: afa0929a943092e4f87a34f8e1c44e23
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
